@@ -15,7 +15,7 @@ import { CdkTableExporterModule } from 'cdk-table-exporter';
 import { InfoService } from 'src/app/services/info/info.service';
 import { Router } from '@angular/router';
 import { BlockUI, NgBlockUI } from 'ng-block-ui';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { LookupService } from 'src/app/services/lookup/lookup.service';
 import { ProductsService } from 'src/app/services/products/products.service';
 import { Chart, registerables } from 'chart.js';
@@ -51,7 +51,7 @@ export class ProductsHomeComponent extends CdkTableExporterModule implements OnI
   newPamphlet: any[] = [];
   newPamphletSKUs: any[] = [];
   showNewPamphletPanel: boolean = false;
-  newProductForm !: FormGroup;
+  newProductForm !: UntypedFormGroup;
   productManagerProducts: Product[] = [];
   productManagerRole: boolean = false;
   loggedInUser: any;
@@ -85,7 +85,7 @@ export class ProductsHomeComponent extends CdkTableExporterModule implements OnI
     public _liveAnnouncer: LiveAnnouncer,
     public info: InfoService,
     public router: Router,
-    public formBuilder : FormBuilder,
+    public formBuilder : UntypedFormBuilder,
     public lookup: LookupService,
     public products: ProductsService
   ) {
